:lipstick: Add base styles
:heavy_plus_sign: Install Normalize CSS
:lipstick: Create sticky footer
:lipstick: Create container class for spacing out content
:lipstick: Create base header styles
:lipstick: Create base footer styles
:lipstick: Create base link styles

Design ideas
I am inspired by a recent trend for brutalism. Derived from architecture, brutalism is when a design exposes the bare materials which it is constructed from. The implementation can range from deeply uncomfortable designs, to a love-letter to the roots of the construction.

In my take, I want to explore what happens to a design if you omit the structure of colored shapes and borders, instead using whitespace, sizing, and positioning of elements to create a look that is still cohesive and elegant.

To this effect, I am initially thinking that I'll intensify link underlines and other default properties, restrict my color palette to a few distinctive and somewhat harsh colors, ensure that alignment of the elements in my design are exact and thoughtful, and use colors and fonts with a personality that reflects web-safe colors and fonts.