:bento: Add High-Contrast Photo
One aspect of my intended design aesthetic is to be black and white, with very deliberate pops of color. I intended for this photo to be black and white, and reduce the gray shades in between appropriately.

It was critical that the background be a perfect white, to get my aesthetic across.

Again, I'm aiming to create an experience that is restrained and brutalist at first glance, but in which each interactable or responsive element moves with a smooth precision.

This may not be the final image I choose, but I hope it serves this purpose.