Implement Photo and Adjust Interface Surrounding it
I've created sensible constraints for the sizing of the photo of myself and utilize object-fit and object-position so the image is scaled accordingly.

Although I love the fluidity in which it moves, my face frankly gets scarier and scarier the wider its viewed on a desktop.

:warning: TODO: Ideally the photo should not zoom in as much. I propose that I crop myself out of the background and create another background, about 3 times as wide, also in black and white.
:bulb: While I'm at it, it may be more pleasant to move myself from the center of the image to perhaps the center-left and set a focus point, so that I scale around that position..
I can also take steps to ensure that no areas of solid black appear at the top or bottom of the image, so there isn't as much of a harsh edge on the top and bottom.

Another adjustment I made was to the container class. It previously didn't have any padding within it. I added padding which only appears when the container's max-width hasn't been reached. I'm not sure that I like deviating from REM units for this, so I may revisit later.

:bulb: NOTE: I need to step out of my comfort zone a little more. I am accusomted to using wide margins and padding, and as part of this aesthetic, I need to be willing to use small amounts of space in a way that delivers more.
Ideally, my content heirarchy should use Type first and foremost: family, weight, size, and style should all be deliberate.
Regarding whitespace, I should aim for all elements to be spaced just-enough that the separation of them is clear. This is so that for an important design element, I can add extra space to deliver an extra impact.
For every situation I add padding or margin to an element, attempt to reduce the whitespace by 1/2 to force myself to become more comfortable with it.

Moving on, I need to collect content, lay it out, and then once I have enough, I can select the type I wish to use.